@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 
 class ReusableCard extends StatelessWidget {
   final Color color;
-  final Widget child;
-  final Function onPressed;
+  final Widget? child;
+  final Function()? onPressed;
 
-  ReusableCard({@required this.color, this.child, this.onPressed});
+  ReusableCard({required this.color, this.child, this.onPressed});
 
   @override
   Widget build(BuildContext context) {
@@ -13,6 +13,7 @@ class ReusableCard extends StatelessWidget {
       onTap: onPressed,
       child: Container(
         margin: EdgeInsets.all(10.0),
+        padding: EdgeInsets.all(10.0),
         decoration: BoxDecoration(
           color: this.color,
           borderRadius: BorderRadius.circular(10.0),
